Fix leftover from #8236 and make BufferReader/Writer support Span (#8408)
* Fix leftover from #8236 and make BufferReader/Writer support Span

- Fix leftover nits from @bzbarsky-apple's review of #8236
- In order to add span support cleanly, added Span support to
  Reader and BufferWriter, and fixed all necessary breakage.

Testing done: pass all unit tests and CASE cert tests

* Fix mbedTLS usage of EcdsaAsn1SignatureToRaw

* Apply review comments from @bzbarsky-apple

* Use some span reassign instead of out_size ref

* Restyled by clang-format

* Remove unnecessary nullptr checks handled by Span::size()

* Improve IntegerToDer test coverage

- Assign to output spans
- Use new span function for validity checks (`is_span_usable`)
- Replace an untested CHIPCert.cpp usage with tested version

* Add is_span_usable() tests

* Restyled by clang-format

* Grammar fix to kick CI

* Commit forgotten removal of obsolete ConvertIntegerRawToDER from CHIPCert.h

* Apply review comments from @mspang

* Fix clang

+++ src/app/decoder.cpp
@@ -53,7 +53,7 @@ uint16_t extractApsFrame(uint8_t * buffer, uint16_t buf_length, EmberApsFrame * 
                                .ReadOctet(&outApsFrame->radius)
                                .StatusCode();
 
-    return err == CHIP_NO_ERROR ? reader.OctetsRead() : 0;
+    return err == CHIP_NO_ERROR ? static_cast<uint16_t>(reader.OctetsRead()) : 0;
 }
 
 void printApsFrame(EmberApsFrame * frame)
